Move functions to common-translation
