+++ src/state/translations-mixin.ts
@@ -13,11 +13,8 @@ import {
 import { translationMetadata } from "../resources/translations-metadata";
 import { Constructor, HomeAssistant } from "../types";
 import { storeState } from "../util/ha-pref-storage";
-import {
-  getLocalLanguage,
-  getTranslation,
-  getUserLocale,
-} from "../util/hass-translation";
+import { getTranslation } from "../util/hass-translation";
+import { getLocalLanguage, getUserLocale } from "../util/common-translation";
 import { HassBaseEl } from "./hass-base-mixin";
 
 declare global {
